Rename "src"-folder to "tracker"
