+++ Makefile
@@ -2,19 +2,19 @@ entries ?= 4
 
 all:
 	@echo "Running Tracker"
-	@python3 -m src
+	@python3 -m tracker
 clean:
 	@echo "Cleaning up Tracker files"
-	@-rm src/files/*
+	@-rm files/*
 log:
 	@echo "Showing Tracker log file"
-	@cat src/files/tracker.log
+	@cat files/tracker.log
 test: clean
 	@echo "Running tests"
 	@python3 -m unittest discover
 setup: clean
 	@echo "Setting up test values"
-	@python3 -m src.setup_test_values_in_console -e ${entries}
+	@python3 -m tracker.setup_test_values_in_console -e ${entries}
 coverage: clean
 	@python3 -m coverage run --omit=/usr/*,*__init__.py -m unittest discover
 	@python3 -m coverage report
