+++ tests/test_console_logger/test_console_logger.py
@@ -17,7 +17,7 @@ import unittest
 
 import click.testing
 
-from src.console_logger.console_logger import *
+from tracker.console_logger.console_logger import *
 
 
 class TestConsoleLogger(unittest.TestCase):
