+++ tests/test_csv/test_csv_handler.py
@@ -15,7 +15,7 @@ import unittest
 import shutil
 import os
 
-from src.csv import CSVHandler
+from tracker.csv import CSVHandler
 from .CSVBaseTestingClass import CSVBaseTestingClass
