+++ tests/test_handler/test_entry_handler/test_base_entry_handler_class.py
@@ -15,8 +15,8 @@ import unittest
 
 import pandas
 
-from src.csv.CSVAttributes import CSVAttributes
-from src.handler.entry_handler.BaseEntryHandlerClass import BaseEntryHandlerClass
+from tracker.csv.CSVAttributes import CSVAttributes
+from tracker.handler.entry_handler.BaseEntryHandlerClass import BaseEntryHandlerClass
 
 from tests.test_csv.CSVBaseTestingClass import CSVBaseTestingClass
