+++ tests/test_handler/test_entry_handler/test_entry_handler.py
@@ -15,8 +15,8 @@ import unittest
 
 import pandas
 
-from src.handler.entry_handler.EntryHandler import EntryHandler
-from src.exceptions.InvalidIDOfDateException import InvalidIDOfDateException
+from tracker.handler.entry_handler.EntryHandler import EntryHandler
+from tracker.exceptions.InvalidIDOfDateException import InvalidIDOfDateException
 
 from tests.test_csv.CSVBaseTestingClass import CSVBaseTestingClass
