+++ tests/test_handler/test_summary_handler/test_week_summary_handler.py
@@ -14,7 +14,7 @@ along with Tracker. If not, see <http://www.gnu.org/licenses/>.
 import datetime
 import unittest
 
-from src.handler.summary_handler.week_summary_handler import WeekSummaryHandler
+from tracker.handler.summary_handler.week_summary_handler import WeekSummaryHandler
 
 from tests.test_csv.CSVBaseTestingClass import CSVBaseTestingClass
