+++ tracker/commands/stop/__init__.py
