+++ tracker/console_logger/ConsoleLogger.py
