+++ tracker/console_logger/console_logger.py
