+++ tracker/handler/summary_handler/summary_handler_interface.py
@@ -11,9 +11,18 @@ You should have received a copy of the GNU General Public License
 along with Tracker. If not, see <http://www.gnu.org/licenses/>.
 """
 
-from src.console_logger.console_logger import info
+import pandas
+
+from tracker.handler.entry_handler.EntryHandler import EntryHandler
+from tracker.handler.entry_handler.GroupedEntryHandler import GroupedEntryHandler
 
 
-def print_list_item(message: str) -> None:
-    style = " - "
-    info(style + message)
+class SummaryHandlerInterface:
+    def __init__(self) -> None:
+        self._grouped_entry_handler = GroupedEntryHandler()
+        self._entry_handler = EntryHandler()
+
+        self._data = None
+
+    def summary(self) -> pandas.DataFrame:  # pragma: no cover
+        pass
